@@ -3,16 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.1.2",
-    "react": "^16.13.1",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.2.0",
+    "@testing-library/user-event": "^14.6.1",
+    "react": "^19.0.0",
     "react-copy-to-clipboard": "^5.0.2",
-    "react-dom": "^16.13.1",
+    "react-dom": "^19.0.0",
     "react-hammerjs": "^1.0.1",
-    "react-responsive": "^8.1.0",
-    "react-scripts": "4.0.1",
-    "styled-components": "^5.1.1"
+    "react-responsive": "^10.0.0",
+    "react-scripts": "5.0.1",
+    "styled-components": "^6.1.14"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -36,6 +36,7 @@
     ]
   },
   "devDependencies": {
-    "babel-plugin-styled-components": "^1.11.1"
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "babel-plugin-styled-components": "^2.1.4"
   }
 }
